% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AgentMemory}
\alias{AgentMemory}
\title{Create a new memory instance for conversation history}
\usage{
AgentMemory(type = c("conversation", "vectorstore"))
}
\arguments{
\item{type}{Type of memory to create ("conversation", "vectorstore")}
}
\value{
A Memory instance ready to use with Agent
}
\description{
Create a new memory instance for conversation history
}
