% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{aigen_Agent}
\alias{aigen_Agent}
\title{Agent Class}
\description{
Agent class for an agent that interacts with language models and tools.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-aigen_Agent-new}{\code{aigen_Agent$new()}}
\item \href{#method-aigen_Agent-get_name}{\code{aigen_Agent$get_name()}}
\item \href{#method-aigen_Agent-set_name}{\code{aigen_Agent$set_name()}}
\item \href{#method-aigen_Agent-chat}{\code{aigen_Agent$chat()}}
\item \href{#method-aigen_Agent-build_prompt}{\code{aigen_Agent$build_prompt()}}
\item \href{#method-aigen_Agent-add_tool}{\code{aigen_Agent$add_tool()}}
\item \href{#method-aigen_Agent-use_tool}{\code{aigen_Agent$use_tool()}}
\item \href{#method-aigen_Agent-list_tools}{\code{aigen_Agent$list_tools()}}
\item \href{#method-aigen_Agent-clear_memory}{\code{aigen_Agent$clear_memory()}}
\item \href{#method-aigen_Agent-get_tool_stats}{\code{aigen_Agent$get_tool_stats()}}
\item \href{#method-aigen_Agent-get_info}{\code{aigen_Agent$get_info()}}
\item \href{#method-aigen_Agent-clone}{\code{aigen_Agent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-new"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new agent with a language model and optional memory and tools
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$new(model, name = "Assistant", memory = NULL, tools = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{An instance of aigen_LanguageModel}

\item{\code{name}}{Character string for agent name}

\item{\code{memory}}{Optional ConversationMemory instance}

\item{\code{tools}}{Named list of functions to use as tools}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get the agent's name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set the agent's name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$set_name(new_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-chat"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-chat}{}}}
\subsection{Method \code{chat()}}{
Process user input and generate a response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$chat(user_input, system_prompt = NULL, context_window = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-build_prompt"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-build_prompt}{}}}
\subsection{Method \code{build_prompt()}}{
Build prompt from conversation history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$build_prompt(system_prompt = NULL, context_window = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-add_tool"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-add_tool}{}}}
\subsection{Method \code{add_tool()}}{
Add a tool to the agent's toolkit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$add_tool(name, fn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-use_tool"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-use_tool}{}}}
\subsection{Method \code{use_tool()}}{
Execute a tool by name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$use_tool(tool_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-list_tools"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-list_tools}{}}}
\subsection{Method \code{list_tools()}}{
List all available tools
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$list_tools()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-clear_memory"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-clear_memory}{}}}
\subsection{Method \code{clear_memory()}}{
Clear the conversation memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$clear_memory()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-get_tool_stats"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-get_tool_stats}{}}}
\subsection{Method \code{get_tool_stats()}}{
Get the usage statistics for tools
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$get_tool_stats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-get_info}{}}}
\subsection{Method \code{get_info()}}{
Get agent information including name and tools
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$get_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_Agent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_Agent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_Agent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
