% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageModel.R
\name{aigen_AzureOpenAIModel}
\alias{aigen_AzureOpenAIModel}
\title{AzureOpenAIModel Class}
\description{
Implements functionality specific to Azure OpenAI models.
}
\section{Super class}{
\code{\link[aigenflow:aigen_LanguageModel]{aigenflow::aigen_LanguageModel}} -> \code{aigen_AzureOpenAIModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-aigen_AzureOpenAIModel-new}{\code{aigen_AzureOpenAIModel$new()}}
\item \href{#method-aigen_AzureOpenAIModel-generate}{\code{aigen_AzureOpenAIModel$generate()}}
\item \href{#method-aigen_AzureOpenAIModel-clone}{\code{aigen_AzureOpenAIModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_AzureOpenAIModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_AzureOpenAIModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_AzureOpenAIModel$new(
  deployment_name,
  resource_name,
  api_version = "2024-02-15-preview",
  api_key = Sys.getenv("AZURE_OPENAI_KEY"),
  max_tokens = 1000,
  temperature = 0.7
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_AzureOpenAIModel-generate"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_AzureOpenAIModel-generate}{}}}
\subsection{Method \code{generate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_AzureOpenAIModel$generate(
  prompt,
  system_message = "You are a helpful assistant."
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_AzureOpenAIModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_AzureOpenAIModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_AzureOpenAIModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
