% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageModel.R
\name{aigen_LanguageModel}
\alias{aigen_LanguageModel}
\title{LanguageModel Class}
\description{
Base class for defining language models.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-aigen_LanguageModel-new}{\code{aigen_LanguageModel$new()}}
\item \href{#method-aigen_LanguageModel-generate}{\code{aigen_LanguageModel$generate()}}
\item \href{#method-aigen_LanguageModel-clone}{\code{aigen_LanguageModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_LanguageModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_LanguageModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_LanguageModel$new(
  model_name,
  api_key,
  max_tokens = 1000,
  temperature = 0.7
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_LanguageModel-generate"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_LanguageModel-generate}{}}}
\subsection{Method \code{generate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_LanguageModel$generate(prompt, system_message = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_LanguageModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_LanguageModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_LanguageModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
