% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageModel.R
\name{aigen_OllamaModel}
\alias{aigen_OllamaModel}
\title{OllamaModel Class}
\description{
Implements functionality for local LLMs using Ollama
}
\section{Super class}{
\code{\link[aigenflow:aigen_LanguageModel]{aigenflow::aigen_LanguageModel}} -> \code{aigen_OllamaModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-aigen_OllamaModel-new}{\code{aigen_OllamaModel$new()}}
\item \href{#method-aigen_OllamaModel-generate}{\code{aigen_OllamaModel$generate()}}
\item \href{#method-aigen_OllamaModel-clone}{\code{aigen_OllamaModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_OllamaModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_OllamaModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_OllamaModel$new(
  model_name,
  base_url = "http://localhost:11434/api",
  max_tokens = 1000,
  temperature = 0.7
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_OllamaModel-generate"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_OllamaModel-generate}{}}}
\subsection{Method \code{generate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_OllamaModel$generate(prompt, system_message = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_OllamaModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_OllamaModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_OllamaModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
