% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageModel.R
\name{aigen_OpenAIModel}
\alias{aigen_OpenAIModel}
\title{OpenAIModel Class}
\description{
Implements functionality specific to OpenAI language models with support for preview models
}
\section{Super class}{
\code{aigenflow::aigen_LanguageModel} -> \code{aigen_OpenAIModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-aigen_OpenAIModel-new}{\code{aigen_OpenAIModel$new()}}
\item \href{#method-aigen_OpenAIModel-generate}{\code{aigen_OpenAIModel$generate()}}
\item \href{#method-aigen_OpenAIModel-clone}{\code{aigen_OpenAIModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_OpenAIModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_OpenAIModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_OpenAIModel$new(
  model_name,
  api_key = Sys.getenv("OPENAI_API_KEY"),
  max_tokens = 1000,
  temperature = 0.7,
  supports_system_messages = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_OpenAIModel-generate"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_OpenAIModel-generate}{}}}
\subsection{Method \code{generate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_OpenAIModel$generate(
  prompt,
  system_message = "You are a helpful assistant."
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aigen_OpenAIModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-aigen_OpenAIModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aigen_OpenAIModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
