% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory.R
\name{ConversationMemory}
\alias{ConversationMemory}
\title{ConversationMemory Class}
\description{
Memory class for store and manage conversation history.
}
\section{Super class}{
\code{\link[aigenflow:Memory]{aigenflow::Memory}} -> \code{ConversationMemory}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_message}{\code{ConversationMemory$add_message()}}
\item \href{#method-get_recent_messages}{\code{ConversationMemory$get_recent_messages()}}
\item \href{#method-clone}{\code{ConversationMemory$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="add_entry">}\href{../../aigenflow/html/Memory.html#method-add_entry}{\code{aigenflow::Memory$add_entry()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="clear_memory">}\href{../../aigenflow/html/Memory.html#method-clear_memory}{\code{aigenflow::Memory$clear_memory()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="get_memory">}\href{../../aigenflow/html/Memory.html#method-get_memory}{\code{aigenflow::Memory$get_memory()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="initialize">}\href{../../aigenflow/html/Memory.html#method-initialize}{\code{aigenflow::Memory$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_message"></a>}}
\if{latex}{\out{\hypertarget{method-add_message}{}}}
\subsection{Method \code{add_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversationMemory$add_message(role, content)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_recent_messages"></a>}}
\if{latex}{\out{\hypertarget{method-get_recent_messages}{}}}
\subsection{Method \code{get_recent_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversationMemory$get_recent_messages(n = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversationMemory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
