% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_valid_models}
\alias{count_valid_models}
\title{Get number of installed Ollama models}
\usage{
count_valid_models(parsed_content)
}
\arguments{
\item{parsed_content}{Parsed API response}
}
\value{
Number of valid models
}
\description{
Get number of installed Ollama models
}
\keyword{internal}
