% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateFlow}
\alias{CreateFlow}
\title{Create an orchestrated flow with agents and workflows}
\usage{
CreateFlow(agents = list(), workflows = list())
}
\arguments{
\item{agents}{Named list of Agent instances or configurations}

\item{workflows}{Named list of workflow definitions}
}
\value{
An Orchestrator instance
}
\description{
Create an orchestrated flow with agents and workflows
}
