% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_system_hardware}
\alias{detect_system_hardware}
\title{Monitor active compute device for Ollama inference (Heuristic)
Detect system hardware configuration}
\usage{
detect_system_hardware()
}
\value{
List containing platform details and available hardware
}
\description{
Monitor active compute device for Ollama inference (Heuristic)
Detect system hardware configuration
}
\keyword{internal}
