% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExecuteFlow}
\alias{ExecuteFlow}
\title{Execute a workflow with orchestrated agents}
\usage{
ExecuteFlow(orchestrator, workflow, input, context = list())
}
\arguments{
\item{orchestrator}{An Orchestrator instance}

\item{workflow}{Name of the workflow to run}

\item{input}{Input data for the workflow}

\item{context}{Optional context data}
}
\value{
Results of the workflow execution
}
\description{
Execute a workflow with orchestrated agents
}
