% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{FileOperationsTool}
\alias{FileOperationsTool}
\title{FileOperationsTool Class}
\description{
FileOperationsTool class for a operations such as reading, writing, and listing files.
}
\section{Super class}{
\code{\link[aigenflow:Tool]{aigenflow::Tool}} -> \code{FileOperationsTool}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileOperationsTool-new}{\code{FileOperationsTool$new()}}
\item \href{#method-FileOperationsTool-run}{\code{FileOperationsTool$run()}}
\item \href{#method-FileOperationsTool-clone}{\code{FileOperationsTool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileOperationsTool-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileOperationsTool-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileOperationsTool$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileOperationsTool-run"></a>}}
\if{latex}{\out{\hypertarget{method-FileOperationsTool-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileOperationsTool$run(operation, path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileOperationsTool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileOperationsTool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileOperationsTool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
