% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_hardware_status}
\alias{format_hardware_status}
\title{Format hardware status for display}
\usage{
format_hardware_status(hw_info, system_info)
}
\arguments{
\item{hw_info}{Hardware info from monitor_inference_hardware}

\item{system_info}{System info from detect_system_hardware}
}
\value{
List containing formatted status strings
}
\description{
Format hardware status for display
}
\keyword{internal}
