% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_analysis_tools}
\alias{get_analysis_tools}
\title{Create analysis tools for agent}
\usage{
get_analysis_tools()
}
\value{
Named list of analysis functions
}
\description{
Creates a list of analysis tools that can be passed to an agent
}
