% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%.aigen_LanguageModel}
\alias{\%>\%.aigen_LanguageModel}
\title{Enable piping for AI models, agents, and data
Make models pipe-friendly}
\usage{
`\%>\%.aigen_LanguageModel`(model, f, ...)
}
\description{
Enable piping for AI models, agents, and data
Make models pipe-friendly
}
