% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{The value to check for \code{NULL}.}

\item{y}{The fallback value if \code{x} is \code{NULL}.}
}
\value{
Returns \code{x} if it is not \code{NULL}, otherwise \code{y}.
}
\description{
Returns \code{y} if \code{x} is \code{NULL}; otherwise returns \code{x}.
}
\details{
Null coalesce operator
}
