% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory.R
\name{LimitedConversationMemory}
\alias{LimitedConversationMemory}
\title{LimitedConversationMemory Class}
\description{
Memory class for storing conversation history with a size limit
}
\section{Super classes}{
\code{\link[aigenflow:Memory]{aigenflow::Memory}} -> \code{\link[aigenflow:ConversationMemory]{aigenflow::ConversationMemory}} -> \code{LimitedConversationMemory}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LimitedConversationMemory-new}{\code{LimitedConversationMemory$new()}}
\item \href{#method-LimitedConversationMemory-add_message}{\code{LimitedConversationMemory$add_message()}}
\item \href{#method-LimitedConversationMemory-clone}{\code{LimitedConversationMemory$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="add_entry"><a href='../../aigenflow/html/Memory.html#method-Memory-add_entry'><code>aigenflow::Memory$add_entry()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="clear_memory"><a href='../../aigenflow/html/Memory.html#method-Memory-clear_memory'><code>aigenflow::Memory$clear_memory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="get_memory"><a href='../../aigenflow/html/Memory.html#method-Memory-get_memory'><code>aigenflow::Memory$get_memory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aigenflow" data-topic="ConversationMemory" data-id="get_recent_messages"><a href='../../aigenflow/html/ConversationMemory.html#method-ConversationMemory-get_recent_messages'><code>aigenflow::ConversationMemory$get_recent_messages()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimitedConversationMemory-new"></a>}}
\if{latex}{\out{\hypertarget{method-LimitedConversationMemory-new}{}}}
\subsection{Method \code{new()}}{
Initialize memory with size limit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimitedConversationMemory$new(max_messages = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_messages}}{Maximum number of messages to retain}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimitedConversationMemory-add_message"></a>}}
\if{latex}{\out{\hypertarget{method-LimitedConversationMemory-add_message}{}}}
\subsection{Method \code{add_message()}}{
Add a message while respecting size limit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimitedConversationMemory$add_message(role, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{The role of the message sender}

\item{\code{content}}{The message content}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimitedConversationMemory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LimitedConversationMemory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimitedConversationMemory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
