% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MistralEndpoint}
\alias{MistralEndpoint}
\title{Create a Mistral endpoint instance}
\usage{
MistralEndpoint(
  endpoint_url,
  api_key = Sys.getenv("AZURE_ML_KEY"),
  max_tokens = 1000,
  temperature = 0.7
)
}
\arguments{
\item{endpoint_url}{The endpoint URL}

\item{api_key}{API key}

\item{max_tokens}{Maximum tokens in response}

\item{temperature}{Sampling temperature}
}
\value{
An AzureMLEndpointModel instance configured for Mistral
}
\description{
Create a Mistral endpoint instance
}
