% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{monitor_inference_hardware}
\alias{monitor_inference_hardware}
\title{Monitor active hardware usage during model inference}
\usage{
monitor_inference_hardware(platform)
}
\arguments{
\item{platform}{Platform type from detect_system_hardware}
}
\value{
List of active hardware components
}
\description{
Monitor active hardware usage during model inference
}
\keyword{internal}
