% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ollama_diagnostics}
\alias{ollama_diagnostics}
\title{Updated diagnostic function for model counting}
\usage{
ollama_diagnostics(base_url = "http://localhost:11434/api", verbose = TRUE)
}
\arguments{
\item{base_url}{Base URL for Ollama API}
}
\value{
Diagnostic information
}
\description{
Updated diagnostic function for model counting
}
