% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ollama_install}
\alias{ollama_install}
\title{Install an Ollama model}
\usage{
ollama_install(
  model_name,
  base_url = "http://localhost:11434/api",
  quiet = FALSE
)
}
\arguments{
\item{model_name}{Name of the model to install}

\item{base_url}{Base URL for Ollama API}

\item{quiet}{Suppress progress messages}
}
\value{
TRUE if successful
}
\description{
Install an Ollama model
}
