% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ollama_remove}
\alias{ollama_remove}
\title{Remove an Ollama model}
\usage{
ollama_remove(model_name, base_url = "http://localhost:11434/api")
}
\arguments{
\item{model_name}{Name of the model to remove}

\item{base_url}{Base URL for Ollama API}
}
\value{
TRUE if successful
}
\description{
Remove an Ollama model
}
