% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RunFlow}
\alias{RunFlow}
\title{Execute a workflow}
\usage{
RunFlow(orchestrator, workflow, input, context = list())
}
\arguments{
\item{orchestrator}{An Orchestrator instance}

\item{workflow}{Name of the workflow to execute}

\item{input}{Input data for the workflow}

\item{context}{Optional context data}
}
\value{
Results of workflow execution
}
\description{
Execute a workflow
}
