% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Workflow}
\alias{Workflow}
\title{Create a workflow definition}
\usage{
Workflow(...)
}
\arguments{
\item{...}{Step configurations created by Step()}
}
\value{
A list containing the workflow definition
}
\description{
Create a workflow definition
}
