% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{Agent}
\alias{Agent}
\title{Agent Class}
\description{
Agent class for an agent that interacts with language models and tools.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Agent$new()}}
\item \href{#method-chat}{\code{Agent$chat()}}
\item \href{#method-build_prompt}{\code{Agent$build_prompt()}}
\item \href{#method-add_tool}{\code{Agent$add_tool()}}
\item \href{#method-use_tool}{\code{Agent$use_tool()}}
\item \href{#method-clear_memory}{\code{Agent$clear_memory()}}
\item \href{#method-clone}{\code{Agent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Agent$new(model, memory = NULL, tools = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chat"></a>}}
\if{latex}{\out{\hypertarget{method-chat}{}}}
\subsection{Method \code{chat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Agent$chat(
  user_input,
  system_prompt = NULL,
  context_window = 5,
  temperature = 0.7
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_prompt"></a>}}
\if{latex}{\out{\hypertarget{method-build_prompt}{}}}
\subsection{Method \code{build_prompt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Agent$build_prompt(system_prompt = NULL, context_window = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_tool"></a>}}
\if{latex}{\out{\hypertarget{method-add_tool}{}}}
\subsection{Method \code{add_tool()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Agent$add_tool(tool)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-use_tool"></a>}}
\if{latex}{\out{\hypertarget{method-use_tool}{}}}
\subsection{Method \code{use_tool()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Agent$use_tool(tool_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_memory"></a>}}
\if{latex}{\out{\hypertarget{method-clear_memory}{}}}
\subsection{Method \code{clear_memory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Agent$clear_memory()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Agent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
