% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{CalculatorTool}
\alias{CalculatorTool}
\title{CalculatorTool Class}
\description{
CalculatorTool class for a simple calculator tool for evaluating mathematical expressions.
}
\section{Super class}{
\code{\link[aigenflow:Tool]{aigenflow::Tool}} -> \code{CalculatorTool}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalculatorTool-new}{\code{CalculatorTool$new()}}
\item \href{#method-CalculatorTool-run}{\code{CalculatorTool$run()}}
\item \href{#method-CalculatorTool-clone}{\code{CalculatorTool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalculatorTool-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalculatorTool-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatorTool$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalculatorTool-run"></a>}}
\if{latex}{\out{\hypertarget{method-CalculatorTool-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatorTool$run(expression)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalculatorTool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalculatorTool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatorTool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
