% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory.R
\name{ConversationMemory}
\alias{ConversationMemory}
\title{ConversationMemory Class}
\description{
Memory class for store and manage conversation history.
}
\section{Super class}{
\code{aigenflow::Memory} -> \code{ConversationMemory}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConversationMemory-add_message}{\code{ConversationMemory$add_message()}}
\item \href{#method-ConversationMemory-get_recent_messages}{\code{ConversationMemory$get_recent_messages()}}
\item \href{#method-ConversationMemory-get_messages_by_role}{\code{ConversationMemory$get_messages_by_role()}}
\item \href{#method-ConversationMemory-clone}{\code{ConversationMemory$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="add_entry"><a href='../../aigenflow/html/Memory.html#method-Memory-add_entry'><code>aigenflow::Memory$add_entry()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="clear_memory"><a href='../../aigenflow/html/Memory.html#method-Memory-clear_memory'><code>aigenflow::Memory$clear_memory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="count"><a href='../../aigenflow/html/Memory.html#method-Memory-count'><code>aigenflow::Memory$count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="get_memory"><a href='../../aigenflow/html/Memory.html#method-Memory-get_memory'><code>aigenflow::Memory$get_memory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aigenflow" data-topic="Memory" data-id="initialize"><a href='../../aigenflow/html/Memory.html#method-Memory-initialize'><code>aigenflow::Memory$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConversationMemory-add_message"></a>}}
\if{latex}{\out{\hypertarget{method-ConversationMemory-add_message}{}}}
\subsection{Method \code{add_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversationMemory$add_message(role, content)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConversationMemory-get_recent_messages"></a>}}
\if{latex}{\out{\hypertarget{method-ConversationMemory-get_recent_messages}{}}}
\subsection{Method \code{get_recent_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversationMemory$get_recent_messages(n = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConversationMemory-get_messages_by_role"></a>}}
\if{latex}{\out{\hypertarget{method-ConversationMemory-get_messages_by_role}{}}}
\subsection{Method \code{get_messages_by_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversationMemory$get_messages_by_role(role, n = Inf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConversationMemory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConversationMemory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversationMemory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
