% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory.R
\name{Memory}
\alias{Memory}
\title{Memory Class}
\description{
Base class for storing and managing memory data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Memory-new}{\code{Memory$new()}}
\item \href{#method-Memory-add_entry}{\code{Memory$add_entry()}}
\item \href{#method-Memory-get_memory}{\code{Memory$get_memory()}}
\item \href{#method-Memory-clear_memory}{\code{Memory$clear_memory()}}
\item \href{#method-Memory-count}{\code{Memory$count()}}
\item \href{#method-Memory-clone}{\code{Memory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Memory-new"></a>}}
\if{latex}{\out{\hypertarget{method-Memory-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Memory$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Memory-add_entry"></a>}}
\if{latex}{\out{\hypertarget{method-Memory-add_entry}{}}}
\subsection{Method \code{add_entry()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Memory$add_entry(entry)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Memory-get_memory"></a>}}
\if{latex}{\out{\hypertarget{method-Memory-get_memory}{}}}
\subsection{Method \code{get_memory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Memory$get_memory()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Memory-clear_memory"></a>}}
\if{latex}{\out{\hypertarget{method-Memory-clear_memory}{}}}
\subsection{Method \code{clear_memory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Memory$clear_memory()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Memory-count"></a>}}
\if{latex}{\out{\hypertarget{method-Memory-count}{}}}
\subsection{Method \code{count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Memory$count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Memory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Memory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Memory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
