% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageModel.R
\name{OpenAIModel}
\alias{OpenAIModel}
\title{OpenAIModel Class}
\description{
Implements functionality specific to OpenAI language models.
}
\section{Super class}{
\code{\link[aigenflow:LanguageModel]{aigenflow::LanguageModel}} -> \code{OpenAIModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OpenAIModel$new()}}
\item \href{#method-generate}{\code{OpenAIModel$generate()}}
\item \href{#method-clone}{\code{OpenAIModel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAIModel$new(
  model_name,
  api_key = Sys.getenv("OPENAI_API_KEY"),
  max_tokens = 1000
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate"></a>}}
\if{latex}{\out{\hypertarget{method-generate}{}}}
\subsection{Method \code{generate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAIModel$generate(messages, temperature = 0.7)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAIModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
