% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step.R
\name{Step}
\alias{Step}
\title{Step Class}
\description{
Step class for a single step in a chain, defined by its name and action.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Step-new}{\code{Step$new()}}
\item \href{#method-Step-execute}{\code{Step$execute()}}
\item \href{#method-Step-clone}{\code{Step$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Step-new"></a>}}
\if{latex}{\out{\hypertarget{method-Step-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Step$new(name, action)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Step-execute"></a>}}
\if{latex}{\out{\hypertarget{method-Step-execute}{}}}
\subsection{Method \code{execute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Step$execute(input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Step-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Step-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Step$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
