% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.aigen_report}
\alias{summary.aigen_report}
\title{Enhanced summary method for aigen_report objects}
\usage{
\method{summary}{aigen_report}(object, ...)
}
\description{
Enhanced summary method for aigen_report objects
}
