% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{Tool}
\alias{Tool}
\title{Tool Class}
\description{
Tool class for a generic structure for implementing tools.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tool-new}{\code{Tool$new()}}
\item \href{#method-Tool-run}{\code{Tool$run()}}
\item \href{#method-Tool-clone}{\code{Tool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tool-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool$new(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool-run"></a>}}
\if{latex}{\out{\hypertarget{method-Tool-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool$run(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
