% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{VisualizationTool}
\alias{VisualizationTool}
\title{VisualizationTool Class}
\description{
VisualizationTool class to create visualizations using the ggplot2 library.
}
\section{Super class}{
\code{\link[aigenflow:Tool]{aigenflow::Tool}} -> \code{VisualizationTool}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VisualizationTool-new}{\code{VisualizationTool$new()}}
\item \href{#method-VisualizationTool-run}{\code{VisualizationTool$run()}}
\item \href{#method-VisualizationTool-clone}{\code{VisualizationTool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizationTool-new"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizationTool-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizationTool$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizationTool-run"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizationTool-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizationTool$run(data, plot_type, x, y = NULL, color = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizationTool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizationTool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizationTool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
